# cmd/app/swagger-gen/swagger.yaml
openapi: 3.0.1
info:
  contact: {}
  description: "REST API for e-commerce (products, users, orders)"
  title: R2 Ecoomerce API
  version: "1.0"
servers:
- url: /v1
paths:
  /auth/login:
    post:
      description: Authenticate and receive a JWT token
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_user_adapters_http.loginRequest"
        description: Login credentials
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: access_token
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
      summary: Login
      tags:
      - Auth
      x-codegen-request-body-name: credentials
  /auth/register:
    post:
      description: Create a new user account
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_user_adapters_http.registerRequest"
        description: User input
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/r2-challenge_internal_user_domain.User"
          description: Created
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Register user
      tags:
      - Auth
      x-codegen-request-body-name: user
...
  /products/{id}:
    delete:
      description: Delete a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "204":
          content:
            application/json:
              schema:
                type: string
          description: No Content
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Delete product
      tags:
      - Products
    get:
      description: Get a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get product
      tags:
      - Products
    put:
      description: Update a product by ID
      parameters:
      - description: Product ID
        in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_product_adapters_http.updateProductRequest"
        description: Product input
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/r2-challenge_internal_product_domain.Product"
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Update product
      tags:
      - Products
      x-codegen-request-body-name: product
...
  /users/{id}:
    get:
      description: Get a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: true
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      summary: Get user
      tags:
      - Users
...
  /users/me:
    put:
      description: Update authenticated user's profile
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/internal_user_adapters_http.updateProfileRequest"
        description: Profile input
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/r2-challenge_internal_user_domain.User"
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      summary: Update my profile
      tags:
      - Users
      x-codegen-request-body-name: body
components:
  schemas:
...
    internal_product_adapters_http.updateProductRequest:
      properties:
        category:
          type: string
        description:
          type: string
        inventory:
          minimum: 0
          type: integer
        name:
          minLength: 3
          type: string
        price_cents:
          minimum: 0
          type: integer
      required:
      - category
      - name
      - price_cents
      type: object
    internal_user_adapters_http.loginRequest:
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
    internal_user_adapters_http.registerRequest:
      properties:
        email:
          type: string
        name:
          minLength: 3
          type: string
        password:
          minLength: 6
          type: string
      required:
      - email
      - name
      - password
      type: object
    internal_user_adapters_http.updateProfileRequest:
      properties:
        email:
          type: string
        name:
          minLength: 3
          type: string
      required:
      - email
      - name
      type: object
...
    r2-challenge_internal_product_domain.Product:
      example:
        updated_at: updated_at
        name: name
        created_at: created_at
        description: description
        price_cents: 0
        id: id
        category: category
        inventory: 0
      properties:
        category:
          type: string
        created_at:
          type: string
        description:
          type: string
        id:
          type: string
        inventory:
          minimum: 0
          type: integer
        name:
          minLength: 3
          type: string
        price_cents:
          minimum: 0
          type: integer
        updated_at:
          type: string
      required:
      - category
      - name
      - price_cents
      type: object
    r2-challenge_internal_user_domain.User:
      example:
        role: role
        updated_at: updated_at
        name: name
        created_at: created_at
        id: id
        email: email
      properties:
        created_at:
          type: string
        email:
          type: string
        id:
          type: string
        name:
          minLength: 3
          type: string
        role:
          type: string
        updated_at:
          type: string
      required:
      - email
      - name
      - role
      type: object
...